Don't bother tracking the compose-discard ajax request.
The compose-discard ajax request only applies to discards from the
compose window, not other places such as the drafts list (or from
non-Streak gmail clients). In Streak, the main thing we do relating to
discarding is try to stop someone from hitting the button to discard a
send-later, which doesn't involve the ajax request anyway.