Fix code for waiting until attachment cards are loaded.
GmailAttachmentAreaView called Promise.all on an array of event streams.

The code tried to wait for all file attachment cards to have their
buttons loaded, but it would recognize file attachment cards by the
presence of buttons, so it would never wait. Recognize them by the
presence of a download url instead.