Fix error message.
I wanted the logic to be that we set the logger to know the user's email
address, and then we wait until the account switcher DOM is ready so we
can avoid useless errors about it later. The method that got the user's
email address would also look up their name, which depended on the
account switcher DOM, so we'd get the error anyway.