Fix stupid gmail focus issues, make compose sizing with status bars more dependable including in not-fullscreen mode.
Before, when Gmail tried to change the compose's size and then we fought
it, it would sometimes trigger a scroll event. For some stupid reason,
if you have the cursor in the recipients field and scroll the body,
Gmail decides to focus the body. (Wtf) Usually that just results in a
stupid user experience, but when our size-fighting code accidentally
triggers scroll events, things go to hell. If you scroll the body down,
then try to click on the recipients area, Gmail tries to resize the
body, we fight it and resize the body again, then the browser decides
that two size changes means a scroll event should happen (even though
both the size changes being applied puts us back at the same size it was
during the last render), and then Gmail decides a scroll event means it
should unfocus the recipients area and focus the body. tl;dr: You scroll
the body, you click recipients, and your cursor ends up back in the body
again.

My work-around is to not manipulate the element's styles directly as
before, but instead create a CSS stylesheet dynamically that adds
!important rules to the element. This simplifies our fight with Gmail
for sizing the element a lot as it makes our changes simply supersede
Gmail's changes. When Gmail tries to resize the compose, its changes
don't actually affect the element now, and no spurious scroll event
happens.