Finally fully fix the compose pop-ups on view change issue!
Instead of setting location.hash to change views, we create a fake
invisible link element, set its href, simulate a click on it, then
remove it.

If the page hash doesn't change, then we fallback to the regular method
of setting location.hash because it must be a custom hash.