Fix missing message view bug.
If you load a thread with a lot of open messages (if you open a large
thread from search, then that can trigger the thread's messages to all
start as expanded), then Gmail loads them into the page over time. We
would add each message view to the thread view's list of messages
immediately, but we would wait for a millisecond before emitting the
message to the platform-implementation and letting it initialize the
message view. If someone called the getMessageViews method on the thread
view during this period, the error would trigger. I removed that
millisecond delay.

I did discover a new bug that when the threadview is first initialized,
all the message views might not be available yet and some will be
missing. Not sure how to address that yet. Applications should use
registerMessageViewHandler to be sure to receive message views rather
than calling getMessageViews() on thread views.

I tried to do the original workaround we planned: make it so the
platform-implementation uses a general function for reading a view out
of the membrane map or lazily creating it as needed, but that ran into
problems because many of the views' constructors required extra
parameters which the membraneGetter (and all of its callers) would need
to supply for specific view types. I gave up on that route as it wasn't
better.