Fix back button loop.
Streak's back buttons in custom views all activate the browser back
functionality.

When you browse from a custom view to a thread, the InboxSDK would
remember the previous custom view, and then override Gmail's back button
to go it. Which would be a new history item.

So if you browsed from a box to an email and then repeatedly pressed the
in-page back buttons, you'd swap back and forth from the email to the
box.

Now when the InboxSDK overrides the Gmail back button, it makes it use
the browser's back functionality too.