Flow-type SectionView and CollapsibleSectionView.
CollapsibleSectionView was inheriting SectionView's methods, but not
calling its constructor or EventEmitter's constructor! It was only
because EventEmitter's constructor does nothing that this didn't blow
up, and it was likely that future changes to it could accidentally
depend on SectionView's constructor. This has now been sorted out so
both extend EventEmitter instead, and in ES6 style where that type of
mistake isn't possible.