Force ButterBar.showLoading() messages to remain until destroyed
Looking at/thinking about the behavior of `ButterBar.showLoading()`
compared to other methods, it kind of seems like the idea that a
loading message remains until explicitly dismissed by calling
`message.destroy()` is a core part of the expected behavior.
I can't think of a valid reason why one might want to call
`showLoading()` with a `time` less than `Infinity` — unless it's for a
purpose that isn't quite a loading message, at which point
`showMessage()` should be preferred. Since the `time` option isn't
documented for `showLoading()` right now, it seems safe to just force
messages created by `showLoading()` to have a `time` of `Infinity`.
Not sure if this should also apply to `showSaving()`, but I'm guessing
it might.