Add docs for unregistering route handlers
The docs UI doesn't show the annotations for the return functions
of `Router.handleXYZ()` functions, even though they are in the source.
At present this means the return type is simply 'function' and it's
not clear that the `handleXYZ()` functions allow you to unregister.

This adds some text in the function description about unregistering,
adds an unregistration example to the `handleListRoute()` example,
and also fixes a couple typos along the way.