Remove useless return type descriptions, add to fn descriptions
We had a bunch of descriptions for return types in the docs which
didn't show up due to the way we process jsdoc files. This removes
all of the old descriptions and adds further context to the doc entries
for the functions they belong to where necessary to improve clarity.