Add 'soft' route ID/param validation, deprecate param string
We'd like to begin validating that route parameters passed into
`Router.goto()` are both the correct parameters for the corresponding
route ID *and* do not contain any extra, unused parameters. Because
this may break existing users we're going to start by attempting to use
the new, more strict `populateRouteID()` function and then fall back to
the old, more lenient approach if that fails (while simultaneously
logging an error). The hope is that we can pinpoint what we're getting
ourselves into by flipping the switch over to `populateRouteID()` and
make an educated call after looking at the error logs.

This change also adds a few ancillary things:

- Tell flow that route parameters can also be numbers (not just
strings), and explicitly coerce to `String` in `populateRouteID()`.
- Modify the parameter matching logic in `populateRouteID()` to
accept parameters which are not immediately preceded by a `/`,
because in Gmail we currently append a weird `p` to list views with
a page parameter.
- In order to avoid tripping up `populateRouteID()`, only add a default
`page` parameter when one is explicitly added to the supplied route ID.
- We want to remove the ability to supply a single parameter to
`Router.goto()` as a string, rather than an object (the ability to pass
a string is not documented). There is currently a PR out to remove
this usage in Streak (StreakYC/MailFoo#3626), but for now we are going
to log a deprecation warning in order to monitor how many other user
(if any) are using this pattern.