Make GmailResponseProcessor test for `start` param more robust
Prior to this were were only checking whether or not a serialize ->
deserialize -> reserialize process retained the same start that was
in the original data. This was satisfactory at the time of writing
because the existing behavior before adding a `start` param was to
*always* default to 0, so simply testing that a non-zero start was
preserved proved that we had made the desired change.

To protect against future breaking changes, however, we really need
to check whether or not we can successfully change the start from its
original value before serialization.