Add support for `hasMore` option, rejigger API slightly
Add a `hasMore` boolean flag which can be used in combination with the
'many' results mode. API looks roughly like this:

- `threads` must always be present.
- `hasMore` is an optional boolean for use in infinite-total mode.
If `true`, it will show 'many' in the UI. If `false`, it will show
a numeric total based on the page offset (aka `start`) or and how many
threads are on the last page.
- `total` must be either a number (for finite-total mode)
or `'MANY'` (for infinite-total mode). The `'MANY'` form is optional
because `hasMore` is always required when using infinite-total mode,
and therefore ignores the value of `total`.