Wire up router code so that custom list routes work past page 1
The code for detecting and managing URL hash changes in relation to
custom list routes was originally designed to have pretty brittle
support for the first page — i.e. after page 1 it essentially didn't
do anything of value. This wires up the routing code so that it:

- Recognizes navigation to subsequent pages and keeps the custom route
hash in sync (including `/p*` parameters on the URL).
- Handles navigation directly to pages in a list route past page 1 via
the URL and fires appropriate handlers.