More graceful error handling
It looks like the custom list view logic was trying to map handler
result validation errors into an empty array (this was when handlers
returned an array) so that the view would still render, but with zero
threads. Instead, it was simply transforming the error output into an
array, meaning that subsequent operators on the stream were still
seeing them as errors, not values. This lead to the UI simply freezing
whenever a validation error occurred.

This change swaps out `mapErrors()` for `flatMapErrors()`, since
the streams spawned by the latter can emit values into the
resulting stream in addition to errors. It also switches
the mapped output to the new object result type.