Add parser for search input field, tests
- Add the `<input>` element that contains search text to the parser
output for the search bar.
- Split out search bar test from app toolbar location test since
the search test is about to get larger. Also remove some unused imports.
- Add a parser test, along with a new HTML sample for the top nav area
that the finder parser are operating in. I didn't add a case with the
existing HTML sample because it appears to have a very different
structure — one that doesn't even support finding the search input
without going in and doing a ton of extra logic (looking for parent
elements, etc.). Need to figure out if that's worth looking into or
irrelevant.