FINALLY finished tests, fix issue w/ AJAX sniffing
- Add fully working integration tests.
- Fix an issue with recognizing outgoing suggestions requests in
AJAX intercept code. Turns out that when the current user is only
signed in to a single Google account, the `/u/<accountNumber>` piece
of the URL path is omitted. Because I had developed the original AJAX
intercept code to monitor outgoing connections with URLs including
that part of the path (given that I was logged into 2 accounts)
I had to rework the regex to make that piece optional.