Fully support getDraftID(), getCurrentDraftID()
- Associate a composeView instance/DOM element with a draft ID.
- Wire up getDraftID() and getCurrentDraftID().

NOTE: Because Inbox locally generates draft IDs prior to the initial
draft save, both getDraftID() and getCurrentDraftID() are resolved
immediately with an ID. I don't currently know of a case where
the promise will resolve to null if the code is working properly.

This also means that the existing Gmail semantic where the absence of a
draft ID implies a draft has not been saved yet is no longer valid when
running in Inbox. I need to make a minor edit to the docs to clarify
this, but at least theoretically it is a pretty serious anti-pattern
to write application code in a way that tries to infer whether or not
a draft has been saved via getDraftID() or getCurrentDraftID().
Users should really be waiting for a destroy event to make the
assumption that a draft save has occurred.