Add 'sending' event support (with some complications)
Turns out Inbox removes the Compose window *before* the send request
goes out — in fact it can sometimes even be a noticeable (near 1 second)
delay between the compose window disappearing and the request going out.
As a result, the way we handled the ComposeView API in Gmail starts to
break down.

In Gmail, the ComposeView stays around until the actual send occurs,
which means that the natural progression is:

1. user hits send
2. 'presending' fires
3. 'sending' fires
4. 'sent' fires
5. Compose window is removed
6. 'destroy' fires

Because Inbox immediately removes the Compose window, without
modifications the 'destroy' is emitted before a 'sending' event can fire
and as a result the ComposeView stops emitting any events ('sending',
'sent', or otherwise). The approach I've taken as of right now is
to essentially force the ComposeView to remain un-destroyed until
'sending' and 'sent' have fired. This has the unfortunate (but largely
unavoidable) consequence of causing application code to see ComposeViews
that have already been removed from the page as 'un-destroyed'.

We may have to revise the way this works, but for now it's
working alright.