*Actually* support disabled buttons
Turns out the initial creation of a ComposeView button didn't *really*
support a disabled state. The update path supports the logic for
disabling, which meant that subsequent updates via a user-supplied
stream would work, but buttons created using a static object would
not respect the `enabled` option nor would the initial creation of
a button from the first value in a stream.

This change adds logic to the creation path to actually use the
`enabled` option, add the necessary disabled class and ignore
keyboard/click events if applicable.

Tested and works as intended.