Handle DOM re-entry case
A few changes to deal with the fact that ComposeViews get detached from
the DOM after hitting send, but may be reattached later if the send
is canceled or fails.

- Keep track of ComposeView drivers which are 'active' (not destroyed)
and don't construct a new driver if a previously used DOM element is
re-entering.
- Emit a 'sendCanceled' event when a ComposeView element re-enters,
which reverses the 'sending' --> 'sent' --> 'destroy' transition
that is currently in progress.