Fix issue w/ custom thread lists (gmail changes)
Turns out the first message in a Gmail thread doesn't *always* have
a message ID that matches a thread ID. The main case where we've seen
this occur is when someone removes the original first message via the
Gmail API (we do this in Streak).

For custom list views, the user supplies thread IDs and we need to turn
them into RFC message IDs. In cases where thread ID != first message ID,
our code was failing because it would try to lookup the RFC ID for
a message that doesn't exist.

To fix, we now take the thread ID given to us by the user-supplied
handler and make a Gmail request to get the thread details (this
includes the message ID). Once we have a mapping from thread ID ->
messageID, we go ahead with our normal flow of fetching that
message ID's RFC ID.