Tolerate one other schema
I can't tell yet whether or not this is actually a message *send*
schema, but it appears sometimes the '14' key is '2' instead, while
the rest of the structure stays the same. The only way to appropriately
tell whether or not this is used for sending (as opposed to just draft
saves), we need to be able to traverse down into the action list
and parse it so we can log the actions and compare against draft
save/sends.