Start listening for send during presend, better req. parsing
- It looks like we might be listening for send requests *too late*
in a minority of cases by waiting until the ComposeView is removed
from the DOM. This changes the `sending`/`sent` logic to start listening
as soon as we get a `presending` event (and turn it back off if we hear
a `sendCanceled`). Then when the ComposeView is removed we can check
whether or not a sending/sent has already fired, and if not wait for it.
- I noticed that if there are multiple items in the update list for an
item sync request we would stop on the first one that was a message
update, compare its actions to the send actions list, and then decide
whether or not we were seeing a send request based on that.
In cases where there is a draft update and then a send update, this
would lead to a match failure even though we could've matched if we had
continued on to the second item. To fix, this moves the send action
comparison into the update list iteration to ensure we find a match
even if it's not the first update in the list. To retain logging of
un-sanitized actions I had to slightly rework how the logging works,
but it seems to behave as expected.