Log first 10 chars of data, remove action logging
Going to start logging the first 10 chars of string properties in the
JSON we're parsing in order to get a better view of what the data
actually looks like without comprising privacy. As a result,
we no longer need to explicitly log actions for requests because
(to my knowledge) actions are never that long.