Queue up 3 failed send parses before reporting
From what I can tell, we're missing out on a small number of outgoing
send requests because they happen before we start listening for them.
I can't tell for sure, but I also have a suspicion that some send
requests are actually in the second request after hitting send on a
ComposeView (perhaps because the first request was already enqueued).
I've attempted to solve the first case by beginning to listen for send
requests as soon as we enter the presending phase. To 1) confirm whether
or not that is working and 2) get sane logging results in a world where
we aren't assuming that the single request directly following a
ComposeView getting removed from the DOM is the only candidate for an
outgoing send, this change takes the following approach:

1) inform the AJAX intercept code that we're presending and prep it
for reporting on failed send parsing.

2) inform the AJAX intercept code in the event of a canceled send during
presending so it can revert to its prior state.

3) rather than only paying attention to the single request directly
following the notification that a send is in progress, enqueue up to
3 outgoing requests which fail to match as a send request. When we hit
3, assume that one of those requests must've been the send request, and
report an array of all 3 so we can take a look in the logs.

4) if less than 3 requests are enqueued, and then we subsequently match
a send request, clear the enqueued requests and reset.