Fix handling of send cancels caused by Gmail
Just uncovered this issue while dealing with events in Inbox, so fixing
on the Gmail side as well.

When something about an email (e.g. no recipients) causes Gmail to
cancel a send, we fire a `presending` event but never recognize that
the send was aborted. Since we already have the `sendCanceled` event,
it's pretty simple to express that behavior via the public API.

This change checks whether or not the compose window is still visible
15ms after a `presending` (to let Gmail's JS handle the send
appropriately), then dispatches a `sendCanceled` if one hasn't been
fired already.