Handle gmail-canceled send better (esp. replies)
While the previous way we were handling this was ok for compose moles,
turns out the inline reply element actually doesn't get hidden until
the email send is complete. The end result is that a `sendCanceled` is
triggered for every email send via inline compose.

Lucky for us, it appears Gmail triggers a sending request synchronously
(ish?), so we can rely on on the `sending` event instead of DOM
visibility to inform us that the email is actually getting sent out to
the server.