Fix sizing of thread row images for rows w/ action links
Now that thread rows inside bundles are getting modified, I've noticed
a visual bug with `addImage()` has appeared due to the image size
being too large. On thread rows with action links (e.g.
"view pull request" on GitHub emails), the addition of a second row
inside the thread exposes the fact that SDK-added images actually
use more than their fair share of the top row. The top row of a thread
is really 20px high, but images were set to display 24px high.
Short of bloating the native thread row height when there is an
action link, I've decided to tweak the styles for images to bound them
at 20x20px so they play nice with the native thread height.

This change means that a lot of the re-use of styles for thread row
images is now impossible, which led to pulling out some styles back
into the gmail.css sheet.