Use extractContents()/insertNode() instead of surroundContents()
Turns out when you use `surroundContents()` to wrap a range, it throws
if the range splits a non-text node (i.e. half of a <div>'s contents
are selected). If you use the `extractContents()`/`insertNode()` combo
it is (basically) functionally equivalent, and will clone the split
range instead of trying to re-use it.

As far as I can tell, this preserves the same behavior in all
replacement cases.