Style and API tweaks for NavMenu
- Add padding for 1st-level NavItems that don't have icons so that
their text lines up with the other NavItems rather than hanging
to the left.
- Change `expanderBackgroundColor` and `expanderForegroundColor` to
`backgroundColor` and `foregroundColor`, making it so that if a
`backgroundColor` is provided for a 1st level NavItem it will *always*
show a colored circle, and then add an expand/collapse arrow if the
item gets children.