Remove outdated CSS rule for compose status bars that breaks things now
When this rule is present and a status bar has ever been added to a compose, then clicking into the recipients area when there are any recipients already causes the composeview to change height. This height change during click sometimes causes the compose to minimize (if the final click lands on the titlebar).